Learning Planner Pro/requirements.txt
altgraph==0.17.4
APScheduler==3.11.0
babel==2.17.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
google-api-core==2.24.2
google-api-python-client==2.169.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
httplib2==0.22.0
idna==3.10
macholib==1.16.3
numpy==2.2.5
oauthlib==3.2.2
packaging==25.0
pandas==2.2.3
plyer==2.1.0
proto-plus==1.26.1
protobuf==6.30.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyinstaller==6.13.0
pyinstaller-hooks-contrib==2025.3
pyparsing==3.2.3
PyQt5==5.15.11
PyQt5-Qt5==5.15.16
PyQt5_sip==12.17.0
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
setuptools==80.1.0
six==1.17.0
SQLAlchemy==2.0.40
tkcalendar==1.6.1
tkTimePicker==2.0.2
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.1.1
urllib3==2.4.0
